Update Project structure
