+++ JoinUs/Controller/Login & Register/LoginViewController.swift
@@ -6,105 +6,13 @@
 //  Copyright © 2020 iosNowon. All rights reserved.
 //
 
+
 import UIKit
-import SnapKit
-import Firebase
-
-class LoginViewController: UIViewController {
-    
-    //MARK:- Properties
-    
-    let backgroundView: UIView = {
-        let theBackgroundView = UIView()
-        theBackgroundView.backgroundColor = .white
-        
-        return theBackgroundView
-    }()
-    
-    let loginLabel: UILabel = {
-        let theLoginLabel = UILabel()
-        theLoginLabel.text = "Login"
-        theLoginLabel.font = UIFont.boldSystemFont(ofSize: 35)
-        theLoginLabel.textColor = UIColor(rgb: Constants.colorHexValue)
-        
-        return theLoginLabel
-    }()
-    
-    let idLabel: UILabel = {
-        let theIdLabel = UILabel()
-        theIdLabel.text = "ID"
-        theIdLabel.font = UIFont.boldSystemFont(ofSize: 10)
-        
-        return theIdLabel
-    }()
-    
-    let idTextField: UITextField = {
-        let theTextField = UITextField()
-        theTextField.placeholder = "   exmaple@gmail.com"
-        theTextField.font = UIFont(name: "System", size: 10.0)
-        theTextField.layer.borderWidth = 2.0
-        theTextField.layer.borderColor = UIColor(rgb: Constants.colorHexValue).cgColor
-        theTextField.borderStyle = .roundedRect
-        
-        return theTextField
-    }()
-    
-    let pwLabel: UILabel = {
-        let thePwLabel = UILabel()
-        thePwLabel.text = "PW"
-        thePwLabel.font = UIFont.boldSystemFont(ofSize: 10)
-        
-        return thePwLabel
-    }()
-    
-    let pwTextField: UITextField = {
-        let theTextField = UITextField()
-        theTextField.placeholder = "   *********"
-        theTextField.font = UIFont(name: "System", size: 10.0)
-        theTextField.layer.borderWidth = 2.0
-        theTextField.layer.borderColor = UIColor(rgb: Constants.colorHexValue).cgColor
-        theTextField.borderStyle = .roundedRect
-        theTextField.isSecureTextEntry = true
-        
-        return theTextField
-    }()
-    
-    let loginButton: UIButton = {
-        let theNextButton = UIButton()
-        theNextButton.setTitle("LOGIN", for: .normal)
-        theNextButton.setTitleColor(UIColor.white, for: .normal)
-        theNextButton.backgroundColor = UIColor(rgb: Constants.colorHexValue)
-        theNextButton.layer.cornerRadius = 5
-        
-        theNextButton.addTarget(self, action: #selector(loginButtonTapped), for: .touchUpInside)
-        return theNextButton
-    }()
-    
-    //MARK:- Methods
-    override func viewDidLoad() {
-        super.viewDidLoad()
-        title = ""
-        //        navigationController?.navigationBar.prefersLargeTitles = true
-        
-        
-        backgroundViewConstraints()
-        registerLabelConstraints()
-        idLabelConstraints()
-        idTextFieldConstraints()
-        pwLabelConstraints()
-        pwTextFieldConstraints()
-        nextButtonConstraints()
-        
-        self.navigationController?.setNavigationBarHidden(false, animated: false)
-    }
-    
-}
-
-
+import FirebaseAuth
 
 extension LoginViewController {
-    // Constraints
     
+    // Constraints
     func backgroundViewConstraints() {
         view.addSubview(backgroundView)
         backgroundView.snp.makeConstraints { (make) in
@@ -119,7 +27,6 @@ extension LoginViewController {
             make.leading.equalTo(self.view).offset(40)
             make.size.equalTo(CGSize(width: 150.0, height: 70.0))
         }
-        
     }
     
     func idLabelConstraints() {
@@ -172,7 +79,6 @@ extension LoginViewController {
     }
     
     // 로그인 버튼 터치시 홈 화면으로 이동
-    
     @objc func loginButtonTapped() {
         print(#function)
