+++ JoinUs/VIew/Login & Register/RegisterView.swift
@@ -1,0 +1,114 @@
+//
+//  RegisterView.swift
+//  JoinUs
+//
+//  Created by Yeojaeng on 2020/06/14.
+//  Copyright © 2020 iosNowon. All rights reserved.
+//
+
+import UIKit
+
+class RegisterViewController: UIViewController, UITextFieldDelegate {
+    
+    //MARK:- Properties
+    
+    // Background View
+    let backgroundView: UIView = {
+        let theBackgroundView = UIView()
+        theBackgroundView.backgroundColor = .white
+        
+        return theBackgroundView
+    }()
+    
+    // Register Label
+    let registerLabel: UILabel = {
+        let theRegisterLabel = UILabel()
+        theRegisterLabel.text = "Register"
+        theRegisterLabel.font = UIFont.boldSystemFont(ofSize: 35)
+        theRegisterLabel.textColor = UIColor(rgb: Constants.colorHexValue)
+        
+        return theRegisterLabel
+    }()
+    
+    // ID Label
+    let idLabel: UILabel = {
+        let theIdLabel = UILabel()
+        theIdLabel.text = "ID"
+        theIdLabel.font = UIFont.boldSystemFont(ofSize: 10)
+        
+        return theIdLabel
+    }()
+    
+    // ID TextField
+    let idTextField: UITextField = {
+        let theTextField = UITextField()
+        theTextField.placeholder = "   exmaple@gmail.com"
+        theTextField.font = UIFont(name: "System", size: 10.0)
+        theTextField.layer.borderWidth = 2.0
+        theTextField.layer.borderColor = UIColor(rgb: Constants.colorHexValue).cgColor
+        theTextField.borderStyle = .roundedRect
+        
+        return theTextField
+    }()
+    
+    // PW Label
+    let pwLabel: UILabel = {
+        let thePwLabel = UILabel()
+        thePwLabel.text = "PW"
+        thePwLabel.font = UIFont.boldSystemFont(ofSize: 10)
+        
+        return thePwLabel
+    }()
+    
+    // PW TextField
+    let pwTextField: UITextField = {
+        let theTextField = UITextField()
+        theTextField.placeholder = "   *********"
+        theTextField.font = UIFont(name: "System", size: 10.0)
+        theTextField.layer.borderWidth = 2.0
+        theTextField.layer.borderColor = UIColor(rgb: Constants.colorHexValue).cgColor
+        theTextField.borderStyle = .roundedRect
+        theTextField.isSecureTextEntry = true
+        
+        return theTextField
+    }()
+    
+    // NEXT Button
+    let nextButton: UIButton = {
+        let theNextButton = UIButton()
+        theNextButton.setTitle("NEXT", for: .normal)
+        theNextButton.setTitleColor(UIColor.white, for: .normal)
+        theNextButton.backgroundColor = UIColor(rgb: Constants.colorHexValue)
+        theNextButton.layer.cornerRadius = 5
+        theNextButton.addTarget(self, action: #selector(nextButtonTapped), for: .touchUpInside)
+        
+        return theNextButton
+    }()
+    
+    
+    //MARK:- Methods
+    
+    override func viewWillAppear(_ animated: Bool) {
+        self.navigationController?.navigationBar.topItem?.title = ""
+    }
+    
+    override func viewDidLoad() {
+        super.viewDidLoad()
+        
+        title = ""
+        
+        backgroundViewConstraints()
+        registerLabelConstraints()
+        idLabelConstraints()
+        idTextFieldConstraints()
+        pwLabelConstraints()
+        pwTextFieldConstraints()
+        nextButtonConstraints()
+        
+        navigationController?.setNavigationBarHidden(false, animated: false)
+        //        navigationController?.navigationBar.prefersLargeTitles = true
+        
+        
+    }
+    
+}
